# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "This is the API documentation for your Laravel project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/www/public",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/user-exams": {
            "get": {
                "tags": [
                    "UserExams"
                ],
                "summary": "List all user exams",
                "operationId": "6b7061977234f6e4245cbf83b2113fcd",
                "responses": {
                    "200": {
                        "description": "List of user exams",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserExam"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "UserExams"
                ],
                "summary": "Create a user exam",
                "operationId": "292108556f90a48f3e2ccb8e484232a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "exam_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "exam_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "started_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-08-06T10:00:00Z"
                                    },
                                    "finished_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-08-06T11:30:00Z"
                                    },
                                    "total_time_seconds": {
                                        "type": "integer",
                                        "example": 4500
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "started"
                                    },
                                    "answers": {
                                        "type": "string",
                                        "example": "Answers in JSON format"
                                    },
                                    "describtion": {
                                        "type": "string",
                                        "example": "Initial attempt"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User exam created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserExam"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user-exams/{id}": {
            "get": {
                "tags": [
                    "UserExams"
                ],
                "summary": "Show a user exam",
                "operationId": "0db418257a3b442391fa8807b6d1e5aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User exam found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserExam"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "UserExams"
                ],
                "summary": "Update a user exam",
                "operationId": "31cf891b531278d92bd86ba179835121",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "started_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "finished_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "total_time_seconds": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "answers": {
                                        "type": "string",
                                        "example": "Answers in JSON format"
                                    },
                                    "describtion": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User exam updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserExam"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "UserExams"
                ],
                "summary": "Delete a user exam",
                "operationId": "502c4c730c4fe7c6c9cd2374ec9dd6c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User exam deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "UserExam deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserExam": {
                "required": [
                    "user_id",
                    "exam_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 12
                    },
                    "exam_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-06T10:00:00Z"
                    },
                    "finished_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-06T11:30:00Z"
                    },
                    "total_time_seconds": {
                        "type": "integer",
                        "example": 5400
                    },
                    "status": {
                        "type": "string",
                        "example": "completed"
                    },
                    "answers": {
                        "type": "string",
                        "example": "Answers in JSON format"
                    },
                    "describtion": {
                        "type": "string",
                        "example": "First attempt"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}